third_party/blink/web_tests/external/wpt/css/css-inline/text-box-trim/parsing/text-box-edge-valid.html

<!DOCTYPE html>
<title>text-box-edge parses valid values according to its grammar</title>
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#text-edges">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_valid_value('text-box-edge', 'auto');
test_valid_value('text-box-edge', 'text');
test_valid_value('text-box-edge', 'ex');
test_valid_value('text-box-edge', 'cap');
test_valid_value('text-box-edge', 'ideographic');
test_valid_value('text-box-edge', 'ideographic-ink');
test_valid_value('text-box-edge', 'text alphabetic');
test_valid_value('text-box-edge', 'ideographic text');
test_valid_value('text-box-edge', 'text text', 'text');
test_valid_value('text-box-edge', 'cap alphabetic');
test_valid_value('text-box-edge', 'cap text', 'cap');
test_valid_value('text-box-edge', 'ex alphabetic');
test_valid_value('text-box-edge', 'ex text', 'ex');
</script>
